Nginx UI integration (#278)
### Related issues

- Addresses #235 

### Summary

This PR adds a Dockerfile to build an nginx-based container to serve the
UI; the v3 deployment stack will deploy this image alongside the rest of
the stack. The built UI bundle is baked into the image for ease of
deployment.

The PR piggybacks off of the existing
`.github/workflows/deploy-frontend.yaml` workflow, since the frontend is
built there. In that workflow, the Dockerfile mentioned previously is
built as
`us-central1-docker.pkg.dev/monarch-initiative/monarch-api/monarch-ui`.
During the build, the UI bundle is copied into the image. The image is
then pushed to the GCP Artifact registry, where it can be pulled by the
v3 stack deployer.

### Checks

- [x] All tests have passed (or issues created for failing tests)