Proxies API requests made to nginx to the API container (#354)
Currently, the nginx container just serves the frontend as a static
bundle. This PR adds proxying of `/openapi.json` and any path starting
with `/v3/` to the API container, with all other requests being handled
by the frontend. Requests to anything starting with `/api` are also
redirected to `/v3/api<rest of request URL>`.

### Related issues

- (no related issues)

### Summary

- changes nginx's configuration to proxy requests intended for the API
to the API container

### Checks

- [ ] All tests have passed (or issues created for failing tests)