Fixed text that links to HPO
The text linking to HPO said "The The Phenomics First Resource"; changed it to "The Human Phenotype Ontology".

Made a few other minor edits for clarity and correctness (e.g., Peter likes it to be called "The Phenopacket schema", not "Phenopackets"). The Phenopackets text could use some additional updating, I think, but I didn't attempt to do that.

I also noticed that there's a "Coming Soon" note lower down on the page under Phenopackets, but Phenopackets are covered higher on the page.