Improve association table search (#268)
Initially, this PR is a little disapointing.

It expands the searched fields to include closure IDs and labels, as
well as evidence fields...the thing is, those fields aren't tokenized,
they're all keywords, which means my use case of finding the one
skeletal phenotype associated with Glaucoma requires putting the exact
HPO term name into the search box, and wrapping it in quotes. (🙄)

So, I'm getting this in place, and I'm going to add additional fields
declarations and copy-field statements for association indexing in
monarch-ingest.