add qualifier fields (#533)
This brings in qualifier as an association slot, along with a full
closure expansion. This will replace qualifiers (and related expansion
fields), but this PR will have both, and then a PR will follow after we
have a new KG build with the qualifier fields.

- Replace deprecated qualifiers field with qualifier
- Cleaning up removed slots that were still used
- Update fixtures, restore qualifiers fields for backwards
compatibility. (To remove after the ingest builds with this schema