Node/Yarn to Bun (#624)
Converts repo from using Node and Yarn to just using Bun, and upgrades
packages (notably Vite -> v5).

This is not absolutely necessary. In my testing, it saves some headache
and time, but it's also possible it could cause a problem down the road.
See the added note in the readme.

I think using Bun just as a package manager is very low risk, so at the
very least we could replace Yarn with Bun. In my testing, starting from
cleared caches, Yarn took 1 minute to install the frontend dependencies,
and Bun took 6 seconds (where simply downloading the packages was the
main bottleneck). This could save us time waiting for frontend tests to
finish, and possibly some $$ if we were to ever go over GitHub Actions
quota.

Using Bun as a Node replacement, again from anecdotal evidence, is also
fairly low risk. They seem to have already covered the general, commonly
used Node functionality, and have tested popular tools like Vite to make
sure they work with Bun. In some edge cases though, Bun might not behave
_exactly_ as Node. But that is a bug in Bun, as they explicitly say that
anything in Node should also work in Bun. The other reason it's low risk
is that it shouldn't be difficult to go back to Node if there is ever a
problem. The biggest risk, I think, is there being a problem, and the
maintainers not being aware that it could be a Bun bug.

Using Bun as a bundler or test runner would probably be a bigger change,
and could cause more problems, in my estimation, so that is left out of
this PR.

Another caveat: Bun's support of Windows is still in progress. They say
about 90% of Bun's test suite passes on Windows.

@kevinschaper If someone has a laptop that does NOT have node or yarn
installed on it, I'd be interested to test to make sure that _just_
installing Bun will work as expected. Theoretically, Bun should be the
only installation requirement for the frontend after this.

---------

Co-authored-by: Glass <glass.ships@outlook.com>