Show association selection as association category (#733)
Change up association table category count outputs to be based on the
association category, rather than counterpart category. This can
potentially be simplified by just generating labels from the association
categories, but for now this PR is going to make a smaller change.
Downstream, the UI needs a minor shift in the "sentence" structure of
the association table heading/controls.