Show all association tables together (#775)
Show all association tables at once on a node page rather than having to
select the association type from an option box. From the initial first
pass, I think the minimum to make this work is going to also include
making the page nav always visible and making sure it has scrollspy
functionality. This may turn into a longer lived PR if we find that we
need to tweak table content before we want to merge.