Rewrite the get_publication script (#832)
This overhauls the script to update publications from Google Scholar.
The previous script worked, but it had some drawbacks, namely that it
required manually editing JSON with updated information. The new script
does not, and it outputs data in a format that will create meaningful
diffs with the existing file format when data is updated.

Changes:

  * Moves from argparse to typer for dealing with CLI arguments.

* Separates fetching data from generating the publications.json file
(this was necessary for development to prevent having to hit Google
Scholar on every change).

* Removes need to filter publications based on year-- Google Scholar is
now taken as the single source of truth for publications.

* Takes into account publications whose years have changed in Google
Scholar versus what has been already included in the publications.json
file.

* Better record matching that ignores differences in non-alphanumeric
characters. Previously, there were several false positives that had to
do with changes in punctuation.

  * Logging to stderr rather than an outputted report.

* Single source of truth for adding links to publications without them
in Google Scholar. (Previously, they had to be input by hand in the
produced JSON file).

---------

Co-authored-by: Kevin Schaper <kevinschaper@gmail.com>