Update typer to 0.12 and change command arguments to Annotated API
This is a rather large commit, but should not change any functionality.

First and foremost, all options and arguments are defined using the
typing.Annotated API introduced in typer 0.9.0, which is the recommended
way to use the library: <https://typer.tiangolo.com/release-notes/#090>

That means that this argument:

  foo: str = typer.Argument("default value", "--foo", "-f")

Becomes this:

  foo = Annotated[
    str,
    typer.Argument("--foo","-f")
  ] = "default value"

The Annotated API was already being used in the CLI, but was silently
failing because we were using an older version of Typer that didn't
support it.

Second, the Google-style argument documentation in docstrings was taken
out, as those field descriptions are generated by typer automatically.

Finally, this commit also removes the typer-cli dependency, as it is
deprecated and its functionality is included in typer itself.