Further collect all monarch-py documentation
This commit does two things to complete the collection of all monarch-py
documentation into a single section in the main documentation. It does
the following:

1. Remove "API Reference" section. The information there is in the
   "Interfaces" and "Implementations" subsections of the monarch-py
   documentation.

2. Moves the "Response Data Model"-- which contains documentation
   generated from LinkML models-- into a "Data Model" subsection in
   monarch-py.

Besides the CLI documentation, the monarch-py documentation is a bit
jumbled and obviously needs some attention and changing. It's not quite
clear coming to the page how someone might use the package as a python
package. That is a broader issue, and not one we need to focus on right
now, since it's not clear that anyone but the Web application actually
uses the monarch-py code.