Replace frequency sort key in API call rather than AssociationsTable
This was done to placate TypeScript. Since
frequency_computed_sortable_float is not a type of Association, adding
it as a key in the generic AssociationTable columns required changing a
ton of code. It also doesn't make sense as a key, since that value is
not returned in the record of an entity-- it's only an implicit sort key
used by Solr.