Move `update_publications` make target outside of `make data` (#957)
This creates a separate `make update_publications` command which is
separate from `make data`. The latter depends on an up-to-date Solr
server to be running, while the former does not. It would be useful for
people who don't continually run and update the Solr backend to be able
to update the publications file.