Drastically simplify AppNodeText component
This commit does two main things.

First, it almost completely dials back the difference between rendering
HTML and SVG. Instead of recreating HTML elements (<b>, <i>, <sup>) with
styled <tspan> elements in SVG, it just includes those elements as HTML
in a <foreignObject> tag.

Part of the reason I did this is that there were strange cross-browser
differences in how superscripted text was being truncated. For some
reason, the code for detecting character positions (see the code that
was removed with `getStartPositionOfChar` and `getCharNumAtPosition`)
stopped working when `dy` was set on a tspan in some browsers.

There was also the issue that the SVG equivalents to the HTML elements
were always sliiightly different.

---

Second, and following from that, I was able to get rid of a lot of the
complexities that went with treating rendering SVG and HTML differently.
The new logic is much more simple: the text of the label is retrieved
via `innerHtml`, then a whitelist of tags is escaped (i.e. replacing
`&lt;` with `<` and `&gt;` with `>`).

The only SVG specific logic (besides wrapping the container in a
<foreignObject>) is replacing <sup> tags with <span> tags styled as
superscript. (Again, because of cross-browser rendering issues, which
are commented in the code).