Use `ruff format` for code formatting (#1032)
### Related issues

- Closes #965

### Summary

Drops `black` as a dependency, in favor of using `ruff format` as a
formatter. On my machine, it's a ~100x speedup.

There's a small change to the makefile too. Instead of using the
`--exit-zero` flag in `ruff check`, the command is prefixed with `-` to
[continue running even after a non-zero exit
code](https://www.gnu.org/software/make/manual/make.html#Errors-in-Recipes).

### Checks

- [x] All tests have passed (or issues created for failing tests)

---------

Co-authored-by: Glass <glass.ships@outlook.com>
Co-authored-by: Glass <glass.ships@protonmail.com>